chore: move profile helper functions to the store
