+++ src/common/Alerts/AlertIncompleteProfile.tsx
@@ -1,13 +1,14 @@
 import { observer } from 'mobx-react-lite';
 import { Banner, InternalLink } from 'oa-components';
 import { useProfileStore } from 'src/stores/Profile/profile.store';
-import { isProfileComplete } from 'src/utils/isProfileComplete';
 import { Flex } from 'theme-ui';
 
 export const AlertIncompleteProfile = observer(() => {
-  const { profile: activeUser } = useProfileStore();
+  const { isComplete } = useProfileStore();
 
-  if (!activeUser || isProfileComplete(activeUser)) return null;
+  if (isComplete !== false) {
+    return null;
+  }
 
   return (
     <InternalLink to="/settings">
